+++ test/fixture/verification/called-on/between/supported.php
@@ -1,5 +1,0 @@
-<?php
-
-$message = 'Requires support for closure binding.';
-
-return $detector->isSupported('closure.bind');