+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next release
 
+- **[BC BREAK]** Removed `calledOn()` ([#197]).
 - **[IMPROVED]** Generator spies and iterable spies are now substituted in
   argument matching and verification ([#193]).
 - **[IMPROVED]** Documentation improvements ([#191], [#192]).
@@ -9,6 +10,7 @@
 [#191]: https://github.com/eloquent/phony/issues/191
 [#192]: https://github.com/eloquent/phony/issues/192
 [#193]: https://github.com/eloquent/phony/issues/193
+[#197]: https://github.com/eloquent/phony/issues/197
 
 ## 0.13.5 (2016-09-20)